+++ src/Service/Ses/src/Result/DeleteSuppressedDestinationResponse.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace AsyncAws\Ses\Result;
+
+use AsyncAws\Core\Result;
+
+/**
+ * An HTTP 200 response if the request succeeds, or an error message if the request fails.
+ */
+class DeleteSuppressedDestinationResponse extends Result
+{
+}